Configure automatically lenses-cli at first login
Automatic configuration first checks if zookeeper is up and running and doesn't run in an
error state. If environment variables for user and passwords have been provided then the
`lense-cli` is auto-configured with these credentials. Same if you pass `WEB_PORT`, `DEBUG`
variables, to connect to API with the provided port and debug the usage respectively.