Add a custom page for 502 bad gateway till Lenses run
A feedback given that people ignore the message in the docker run logs
and when they open localhost:3030 page they see a 502 bad gateway which
makes them to understand that Lenses failed to run.

Custom error page with autorefresh every 5s for 502. May later we need
to have a status dashboard to show that all services are up and running.

Issue: FEED-68