Use auto-provisioning (autoconfigure) for Lenses 5.0
This commit adds support to setup Lenses 5.0 via API as setting up connections
via 'lenses.conf' is not supported anymore. Furthermore it adjusts scripts that
rely on Lenses detection to wait for both Lenses to come up and be setup.

Due the change in Lenses setup, we made the delays between chained commands
smaller and we should monitor for failed runs (aka it runs fine on my laptop).

Some settings for old Lenses were adjusted to now reflect the arrays we use in
lenses-cli provision.yaml. So we still use them but instead of adding them
to 'lenses.conf' we add them to 'provision.yaml'. Their format was thus changed
to be an array of URLs, instead of the more detailed old version. These arrays
are comma separated:

   LENSES_ZOOKEEPER_HOSTS="[{url:\"0.0.0.0:$ZK_PORT\",jmx:\"0.0.0.0:$ZK_JMX_PORT\"}]"
   became
   LENSES_ZOOKEEPER_HOSTS=[${LENSES_ZOOKEEPER_HOSTS:-0.0.0.0:$ZK_PORT}]

   LENSES_KAFKA_BROKERS=${LENSES_KAFKA_BROKERS:-"PLAINTEXT://0.0.0.0:$BROKER_PORT"}
   became
   LENSES_KAFKA_BROKERS=[${LENSES_KAFKA_BROKERS:-PLAINTEXT://0.0.0.0:$BROKER_PORT}]

   LENSES_SCHEMA_REGISTRY_URLS=[{url:\"http://0.0.0.0:$REGISTRY_PORT\",jmx:\"0.0.0.0:$REGISTRY_JMX_PORT\"}]
   became
   LENSES_SCHEMA_REGISTRY_URLS=[${LENSES_SCHEMA_REGISTRY_URLS:-http://0.0.0.0:$REGISTRY_PORT}]

   LENSES_KAFKA_CONNECT_CLUSTERS="[{name:\"dev\",urls:[{url:\"http://0.0.0.0:$CONNECT_PORT\",jmx:\"0.0.0.0:$CONNECT_JMX_PORT\"}],statuses:\"connect-statuses\",configs:\"connect-configs\",offsets:\"connect-offsets\",aes256.key:\"0123456789abcdef0123456789abcdef\"}]"
   became
   LENSES_KAFKA_CONNECT_CLUSTERS=[${LENSES_KAFKA_CONNECT_CLUSTERS:-http://0.0.0.0:$CONNECT_PORT}]

The license file now goes under the provision directory.

People can append things to 'provision.yaml' by mounting the '/provision.yaml'
file. Due to the way yaml parser works, the last instance of a key will be used,
so for example if the user provided 'provision.yaml' has the key 'connections'
it will override all autoconfigured connections.

[OPS-1984]

Signed-off-by: Marios Andreopoulos <opensource@andmarios.com>