Add Official arm64 support
This commit sets the build hook to use buildkit, and build for both linux/amd64
and linux/amd64 platforms. It also moves all logic for tags from the post_build
hook to the build hook, to cater for buildx requirements.

It also tweaks arm64 support in the Dockerfile by leveraging docker's env vars
for multi-arch builds, and switching to prebuilt binaries, making the image a
little more “standard”.

[OPS-2157][OPS-2163][OPS-2164]

Signed-off-by: Marios Andreopoulos <opensource@andmarios.com>