Enables kafka-connect REST API to listen on specified port
By default the Kafka Connect REST API listens on port 8083. According to
the documentation, the port should be configurable by specifying
CONNECT_REST_PORT, but for a reason I could not identify this
environment variable is not respected by the running process. The server
process that runs for Kafka Connect will only be available on port 8083
regardless of this environment variable.

Attempting to set CONNECT_PORT or CONNECT_REST_PORT in the container
environment variables causes all dependent services and HTTP connections
to point to the wrong port, because these processes respect the
environment variable, but the actual service continues to run on port
8083.

After digging through online documentation I found the configuration
property that specifies the REST API listeners.

https://docs.confluent.io/platform/current/connect/security.html#connect-rest-api-http

This commit appends the listener interface and port to the
connect-avro-distributed.properties file.